added new file structure and moved navbar related stuff to its own folder
