--- src/Components/Content/Link.js
@@ -1,6 +1,0 @@
-import '../../App.css';
-
-export const Link = (linkData) => {
-    const {id, href, className, text} = linkData;
-    return <a key={id} href={href} className={className}>{text}</a>
-}
